successfully creates fulfillment. checks if buyer exists, checks if buyer has enough currency to buy.

+++ server/gamedata/traderequest.go
@@ -20,7 +20,7 @@ type TradeRequest struct {
 }
 
 type Demands struct {
-	Currency int32 `json:"currency"`
+	Currency float64 `json:"currency"`
 	// TODO: optionally, could have an []Items to allow users to request money + item(s)
 	//		 NO CANCELING TRADE REQUESTS AT THIS MOMENT. This will be one of the later features if time allows
 }
